+++ src/main/java/dev/patika/Vet/App/business/abs/ICustomerService.java
@@ -6,12 +6,12 @@ import dev.patika.Vet.App.entity.Customer;
 import java.util.List;
 
 public interface ICustomerService {
-    Customer getByID(int id);
+    Customer getByID(Long id);
     Customer save(Customer customer);
-    String delete(int id);
+    String delete(Long id);
     Customer update(Customer customer);
     List<Customer> findAll();
     Customer findByName(String name);
 
-    List<Animal> findAnimalByCustomerID(int id);
+    List<Animal> findAnimalByCustomerID(Long id);
 }
